fix: fix parser logic
- changed due to scability
