--- src/main/java/store/service/parser/Parser.java
@@ -1,5 +1,0 @@
-package store.service.parser;
-
-public interface Parser<T> {
-    T parse(String text);
-}
